reconciliation-system/reconciliation-database/src/main/java/com/progressoft/jip11/recdb/RecDbInitializer.java
package com.progressoft.jip11.recdb;

import liquibase.Contexts;
import liquibase.LabelExpression;
import liquibase.Liquibase;
import liquibase.database.Database;
import liquibase.database.DatabaseFactory;
import liquibase.database.jvm.JdbcConnection;
import liquibase.exception.LiquibaseException;
import liquibase.resource.ClassLoaderResourceAccessor;

import javax.sql.DataSource;
import java.sql.Connection;
import java.sql.SQLException;

public class RecDbInitializer {

    public void initialize(DataSource dataSource) {
        validate(dataSource);
        try (Connection connection = dataSource.getConnection()) {
            Database database = DatabaseFactory.getInstance().findCorrectDatabaseImplementation(new JdbcConnection(connection));
            Liquibase liquibase = new liquibase.Liquibase("rec-db.xml", new ClassLoaderResourceAccessor(), database);
            liquibase.update(new Contexts(), new LabelExpression());
        } catch (SQLException | LiquibaseException e) {
            throw new IllegalStateException("error occurred while initializing", e);
        }
    }

    private void validate(DataSource dataSource) {
        if (dataSource == null)
            throw new NullPointerException("datasource is null");
    }
}
